ray drawing for minimap fixed but sometimes clips through walls
