+++ Makefile
@@ -9,7 +9,7 @@ SRCS = main.c ./srcs/parsing/file_type.c ./srcs/utils/init.c ./srcs/utils/parsin
 		./srcs/game_funcs/get_texture.c ./srcs/game_funcs/fill_map.c ./srcs/utils/casting_utils.c \
 		./srcs/utils/texture_utils.c ./srcs/game_funcs/draw_line.c ./srcs/raycaster/caster.c ./srcs/raycaster/horizontal.c \
 		./srcs/raycaster/vertical.c ./srcs/utils/init_utils.c ./srcs/utils/hook_utils.c ./srcs/utils/render_utils.c \
-		./srcs/utils/minimap_utils.c
+		./srcs/raycaster/minimap.c ./srcs/player_funcs/player_movement.c ./srcs/player_funcs/player_rotation.c
 
 OBJS = $(SRCS:.c=.o)
