+++ srcs/game_funcs/game_hooks.c
@@ -3,50 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   game_hooks.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: skunert <skunert@student.42heilbronn.de    +#+  +:+       +#+        */
+/*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/06 13:48:12 by skunert           #+#    #+#             */
-/*   Updated: 2023/10/16 18:59:33 by skunert          ###   ########.fr       */
+/*   Updated: 2023/10/18 21:32:30 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
-
-void	ft_rotate_left(t_game *game, int diff)
-{
-	if (mlx_is_key_down(game->mlx, MLX_KEY_LEFT))
-	{
-		game->caster->pa -= 0.05;
-		if (game->caster->pa < 0)
-			game->caster->pa += 2 * M_PI;
-	}
-	else if (diff > 0)
-	{
-		if ((diff) > 150)
-			diff = 150;
-		game->caster->pa -= 0.0006 * diff;
-		if (game->caster->pa < 0)
-			game->caster->pa += 2 * M_PI;
-	}
-}
-
-void	ft_rotate_right(t_game *game, int diff)
-{
-	if (mlx_is_key_down(game->mlx, MLX_KEY_RIGHT))
-	{
-		game->caster->pa += 0.05;
-		if (game->caster->pa > 2 * M_PI)
-			game->caster->pa -= 2 * M_PI;
-	}
-	else if (diff < 0)
-	{
-		if ((-diff) > 150)
-			diff = -150;
-		game->caster->pa += 0.0006 * (-diff);
-		if (game->caster->pa > 2 * M_PI)
-			game->caster->pa -= 2 * M_PI;
-	}
-}
 
 void	ft_wall_offset_set(t_game *g)
 {
@@ -70,15 +34,6 @@ void	ft_wall_offset_set(t_game *g)
 		g->caster->y_off_strafe = -20;
 	else
 		g->caster->y_off_strafe = 20;
-}
-
-void	ft_move(t_game *game)
-{
-	ft_wall_offset_set(game);
-	if (mlx_is_key_down(game->mlx, MLX_KEY_ESCAPE))
-		mlx_close_window(game->mlx);
-	ft_move_up_down(game);
-	ft_move_left_right(game);
 }
 
 void	ft_hooks(void *param)
