+++ srcs/player_funcs/player_rotation.c
@@ -1,0 +1,50 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   player_rotation.c                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/10/18 21:29:58 by njantsch          #+#    #+#             */
+/*   Updated: 2023/10/18 21:30:59 by njantsch         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../includes/cub3d.h"
+
+void	ft_rotate_left(t_game *game, int diff)
+{
+	if (mlx_is_key_down(game->mlx, MLX_KEY_LEFT))
+	{
+		game->caster->pa -= 0.05;
+		if (game->caster->pa < 0)
+			game->caster->pa += 2 * M_PI;
+	}
+	else if (diff > 0)
+	{
+		if ((diff) > 150)
+			diff = 150;
+		game->caster->pa -= 0.0006 * diff;
+		if (game->caster->pa < 0)
+			game->caster->pa += 2 * M_PI;
+	}
+}
+
+void	ft_rotate_right(t_game *game, int diff)
+{
+	if (mlx_is_key_down(game->mlx, MLX_KEY_RIGHT))
+	{
+		game->caster->pa += 0.05;
+		if (game->caster->pa > 2 * M_PI)
+			game->caster->pa -= 2 * M_PI;
+	}
+	else if (diff < 0)
+	{
+		if ((-diff) > 150)
+			diff = -150;
+		game->caster->pa += 0.0006 * (-diff);
+		if (game->caster->pa > 2 * M_PI)
+			game->caster->pa -= 2 * M_PI;
+	}
+}
+
