+++ srcs/raycaster/horizontal.c
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 21:26:51 by njantsch          #+#    #+#             */
-/*   Updated: 2023/10/18 13:42:29 by njantsch         ###   ########.fr       */
+/*   Updated: 2023/10/18 20:40:56 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
