+++ srcs/raycaster/minimap.c
@@ -1,34 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   minimap_utils.c                                    :+:      :+:    :+:   */
+/*   minimap.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: skunert <skunert@student.42heilbronn.de    +#+  +:+       +#+        */
+/*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/17 14:58:26 by skunert           #+#    #+#             */
-/*   Updated: 2023/10/18 14:53:11 by skunert          ###   ########.fr       */
+/*   Updated: 2023/10/18 21:19:41 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
 
-
-void	ft_fill_minispace(t_game *g, int start_x, int start_y, uint32_t color)
+void	draw_minimap(t_game *g)
 {
-	int	x;
-	int	y;
-
-	y = start_y;
-	while (y < start_y + 20)
-	{
-		x = start_x;
-		while (x < start_x + 20)
-		{
-			mlx_put_pixel(g->minimap, x, y, color);
-			x++;
-		}
-		y++;
-	}
+	g->dl->begin_x = 98;
+	g->dl->begin_y = 98;
+	g->dl->end_x = (g->ray->ray_x - (g->pl_x - 5 * DIMENS)) / DIMENS * 20;
+	g->dl->end_y = (g->ray->ray_y - (g->pl_y - 5 * DIMENS)) / DIMENS * 20;
+	g->dl->end_x = fmin(fmax(g->dl->end_x, 0), 198);
+	g->dl->end_y = fmin(fmax(g->dl->end_y, 0), 198);
+	ft_draw_line(g);
 }
 
 void	ft_fill_miniplayer(t_game *g, int x, int y)
